Move time.ts to /common
