+++ extensions/ql-vscode/src/common/time.ts
