+++ extensions/ql-vscode/src/query-history/query-history-manager.ts
@@ -19,7 +19,7 @@ import {
 } from "../common/vscode/dialog";
 import { URLSearchParams } from "url";
 import { DisposableObject } from "../common/disposable-object";
-import { ONE_HOUR_IN_MS, TWO_HOURS_IN_MS } from "../pure/time";
+import { ONE_HOUR_IN_MS, TWO_HOURS_IN_MS } from "../common/time";
 import { assertNever, getErrorMessage } from "../common/helpers-pure";
 import { CompletedLocalQueryInfo, LocalQueryInfo } from "../query-results";
 import {
