+++ extensions/ql-vscode/src/variant-analysis/variant-analysis-monitor.ts
@@ -11,7 +11,7 @@ import {
 import { VariantAnalysis as ApiVariantAnalysis } from "./gh-api/variant-analysis";
 import { processUpdatedVariantAnalysis } from "./variant-analysis-processor";
 import { DisposableObject } from "../common/disposable-object";
-import { sleep } from "../pure/time";
+import { sleep } from "../common/time";
 import { getErrorMessage } from "../common/helpers-pure";
 import { App } from "../common/app";
 import { showAndLogWarningMessage } from "../common/logging";
