+++ extensions/ql-vscode/src/view/common/LastUpdated.tsx
@@ -3,7 +3,7 @@ import { useMemo } from "react";
 import styled from "styled-components";
 
 import { parseDate } from "../../common/date";
-import { humanizeRelativeTime } from "../../pure/time";
+import { humanizeRelativeTime } from "../../common/time";
 
 import { Codicon } from "./icon";
