+++ extensions/ql-vscode/src/view/variant-analysis/VariantAnalysisStats.tsx
@@ -4,7 +4,7 @@ import styled from "styled-components";
 import { VariantAnalysisStatus } from "../../variant-analysis/shared/variant-analysis";
 import { StatItem } from "./StatItem";
 import { formatDecimal } from "../../common/number";
-import { humanizeUnit } from "../../pure/time";
+import { humanizeUnit } from "../../common/time";
 import { VariantAnalysisRepositoriesStats } from "./VariantAnalysisRepositoriesStats";
 import { VariantAnalysisStatusStats } from "./VariantAnalysisStatusStats";
